dva_ex2_Julian_Kalis_18450494/requirements.txt
numpy
pillow
bokeh
scikit-learn